pyproject.toml
[tool.poetry]
name = "sugatan_shopify"
version = "0.1.0"
description = ""
authors = ["HM <hieumdd@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.9"
requests = "^2.26.0"
google-cloud-bigquery = "^2.30.1"
Jinja2 = "^3.0.3"
google-cloud-pubsub = "^2.9.0"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
pytest = "^6.2.5"
pytest-timeout = "^2.0.1"
mypy = "^0.910"
types-requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
